skills/coco-ppt/src/generator/assembler.py
"""
幻灯片组装器 - 根据匹配结果组装幻灯片
"""

import logging
from typing import List, Optional
from copy import deepcopy
from pptx import Presentation
from pptx.slide import Slide
from ..analyzer.matcher import LayoutMatch

logger = logging.getLogger(__name__)


class SlideAssembler:
    """组装幻灯片"""
    
    def __init__(self, template: Presentation):
        """
        初始化组装器
        
        Args:
            template: 模板 Presentation 对象
        """
        self.template = template
        self.output: Optional[Presentation] = None
    
    def assemble(self, matches: List[LayoutMatch]) -> Presentation:
        """
        根据匹配结果组装幻灯片
        
        Args:
            matches: 匹配结果列表
            
        Returns:
            组装后的 Presentation 对象
        """
        logger.info(f"开始组装幻灯片：{len(matches)}张")
        
        # 创建新的演示文稿
        self.output = Presentation()
        
        # 移除默认幻灯片
        while len(self.output.slides) > 0:
            self.output.slides._sldIdLst.remove(self.output.slides[0]._element)
        
        # 根据匹配复制幻灯片
        for match in matches:
            source_slide = self.template.slides[match.matched_layout.slide_index]
            new_slide = self._duplicate_slide(source_slide)
            logger.debug(f"复制幻灯片 {match.matched_layout.slide_index} -> {len(self.output.slides)}")
        
        logger.info(f"组装完成：{len(self.output.slides)}张幻灯片")
        return self.output
    
    def _duplicate_slide(self, source: Slide) -> Slide:
        """
        复制幻灯片
        
        Args:
            source: 源幻灯片
            
        Returns:
            新幻灯片
        """
        # 获取源幻灯片的版式
        slide_layout = source.slide_layout
        
        # 在输出演示文稿中添加相同版式的幻灯片
        new_slide = self.output.slides.add_slide(slide_layout)
        
        # 复制形状（保留格式）
        self._copy_shapes(source, new_slide)
        
        return new_slide
    
    def _copy_shapes(self, source: Slide, target: Slide):
        """
        复制形状到目标幻灯片
        
        Args:
            source: 源幻灯片
            target: 目标幻灯片
        """
        # 注意：python-pptx 不直接支持形状复制
        # 这里我们依赖版式来保持形状结构
        # 实际的内容填充由 ContentApplier 完成
        
        # 复制背景
        if source.background:
            try:
                target.background = source.background
            except Exception as e:
                logger.warning(f"无法复制背景：{e}")
        
        # 复制备注
        if source.has_notes_slide:
            try:
                source_notes = source.notes_slide
                target_notes = target.notes_slide
                if hasattr(source_notes, 'text') and source_notes.text:
                    target_notes.text = source_notes.text
            except Exception as e:
                logger.warning(f"无法复制备注：{e}")
    
    def reorder_slides(self, order: List[int]) -> Presentation:
        """
        重新排序幻灯片
        
        Args:
            order: 新的顺序（幻灯片索引列表）
            
        Returns:
            重新排序后的 Presentation 对象
        """
        if not self.output:
            raise ValueError("请先调用 assemble() 方法")
        
        logger.info(f"重新排序幻灯片：{order}")
        
        # 创建新的演示文稿
        reordered = Presentation()
        
        # 移除默认幻灯片
        while len(reordered.slides) > 0:
            reordered.slides._sldIdLst.remove(reordered.slides[0]._element)
        
        # 按新顺序复制幻灯片
        for idx in order:
            if 0 <= idx < len(self.output.slides):
                source = self.output.slides[idx]
                # 复制幻灯片
                new_slide = reordered.slides.add_slide(source.slide_layout)
                # 这里应该复制内容，但简化处理
        
        self.output = reordered
        return self.output
    
    def save(self, output_path: str):
        """
        保存演示文稿
        
        Args:
            output_path: 输出文件路径
        """
        if not self.output:
            raise ValueError("没有可保存的演示文稿")
        
        logger.info(f"保存演示文稿：{output_path}")
        self.output.save(output_path)
        logger.info("保存完成")
    
    def get_slide_count(self) -> int:
        """获取幻灯片数量"""
        return len(self.output.slides) if self.output else 0
