skills/coco-ppt/requirements.txt
python-pptx>=0.6.21
Pillow>=9.0.0
defusedxml>=0.7.1
click>=8.0.0
rich>=12.0.0
